pyproject.toml
[project]
name = "olxproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "django==5.2",
    "django-filter>=25.2",
    "django-humanize>=0.1.2",
    "django-jazzmin==2.6.2",
    "django-json-widget>=2.1.1",
    "django-mptt>=0.18.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-recursive>=0.1.2",
    "djangorestframework-stubs>=3.16.8",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.29.0",
    "markdown>=3.10.2",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
]
